[해결] 비트마스킹을 32비트로 하고있었음...
입력은 최대 50개

+++ boj/03018.cpp
@@ -23,7 +23,7 @@ int main() {
         info[number_list[j]] = info[number_list[j]] | 1LL << i;
       }
     } else {
-      int attendance = 0;
+      unsigned long long int attendance = 0;
       for (int j = 0; j < p; j++) {
         attendance = attendance | info[number_list[j]];
       }
